project/requirements.txt
Babel==2.3.4
Django==1.8
Jinja2==2.9.5
MarkupSafe==0.23
PyJWT==1.0.0
PyYAML==3.12
Pygments==2.2.0
Sphinx==1.5.2
WTForms==2.0.2
alabaster==0.7.9
amqp==1.4.6
anyjson==0.3.3
argh==0.26.2
arrow==0.5.4
backports.ssl-match-hostname==3.4.0.2
billiard==3.3.0.20
celery==3.1.17
cent==0.3.0
centrifuge==0.7.0
certifi==2015.04.28
click==6.7
dj-database-url==0.3.0
dj-static==0.0.6
django-allauth==0.19.1
django-bootstrap==0.2.4
django-bootstrap3==5.2.0
django-casper==0.0.4
django-celery==3.1.16
django-debug-toolbar==1.3.0
django-dotenv==1.4.1
django-mailgun==0.2.2
django-node==2.3.3
django-nopassword==1.3.1
django-push-notifications==1.4.1
django-qsstats-magic==0.7.2
django-react==0.9.0
django-template-email==0.1
django-toolbelt==0.0.1
django-uuidfield==0.5.0
django-webpack==2.1.2
djangorestframework==3.1.0
djangorestframework-jwt==1.4.0
docutils==0.13.1
drf-extensions==0.2.7
fluent-logger==0.4.0
graphviz==0.5.1
gunicorn==19.3.0
hiredis==0.2.0
imagesize==0.7.1
jsonfield==1.0.3
jsonschema==2.4.0
kombu==3.0.24
livereload==2.5.1
markdown2==2.3.0
meld3==1.0.2
msgpack-python==0.4.6
oauthlib==0.7.2
opbeat==3.0.2
pathtools==0.1.2
port-for==0.3.1
psycopg2==2.6
python-dateutil==2.4.2
python-dotenv==0.6.3
python-openid==2.2.5
pytz==2015.2
pywatchman==1.2.0
redis==2.10.3
requests==2.5.3
requests-oauthlib==0.4.2
six==1.9.0
snowballstemmer==1.2.1
sockjs-tornado==1.0.1
sphinx-autobuild==0.6.0
sqlparse==0.1.14
static3==0.5.1
toredis-fork==0.1.4
tornado==4.1
toro==0.8
watchdog==0.8.3
wsgiref==0.1.2
graphviz
python-dotenv
